refactor: move file I/O and lock platform code into src/platform
file_writer.cc and cross_process_file_lock.cc each carried Win32/POSIX
implementations behind in-file #ifdef _WIN32 branches. Follow the existing
src/platform/{windows,posix}/path.cc pattern instead:

- Add src/platform/file_io.h declaring fl::platform OpenWritableFile /
  WriteFileAt / CloseFile primitives over an integer handle, implemented per
  OS in src/platform/{windows,posix}/file_io.cc. FileWriter now stores an
  intptr_t handle and delegates, dropping its #ifdef member and branches.
- Split CrossProcessFileLock: the platform-specific State, its releasing
  destructor, FormatProcessInfo, the constructor/destructor, and
  TryAcquireForDirectory move verbatim into
  src/platform/{windows,posix}/cross_process_file_lock.cc; the neutral
  src/download/cross_process_file_lock.cc keeps only the cross-platform
  WaitForDirectoryLock orchestration.
- Wire the four new sources into FOUNDRY_LOCAL_PLATFORM_SOURCES.

No behavior change; addresses Scott's #793 nit about keeping platform code in
src/platform. Verified: Windows build green, 83 download/lock/file-writer/state
tests pass, POSIX sources syntax-clean under g++ -Wall -Wextra.
